Added optimisation settings to all cmake files

+++ y2015cpp/day3/include/task1.hpp
@@ -7,6 +7,6 @@ using namespace std;
 using dir_func = void (*)();
 
 
-void perfTaskOne(char*, size_t, int*);
+void perfTaskOne(char*, int, int*);
 
 #endif
