+++ y2015cpp/day3/src/task2.cpp
@@ -59,7 +59,7 @@ unordered_map<char, dir_func> robo_directions{
 };
 
 
-void perfTaskTwo(char* instructions, size_t size, int* srhouses){
+void perfTaskTwo(char* instructions, int size, int* srhouses){
     unordered_map<int, unordered_set<int>>* rows = new unordered_map<int, unordered_set<int>>{}; // container for data on rows
 
     unordered_set<int> first_row = {0}; // add first point (house) that is visited
